requirements.txt
# Core
python-dotenv==1.0.0
sqlalchemy==2.0.23

# APIs de IA
google-generativeai==0.3.0
openai==1.3.0
ollama==0.1.0

# Web Framework
gradio==4.20.0

# Utilities
requests==2.31.0